- Renamed `ExtensionsSystem.cs` to `ExtensionsGeneral.cs` - Removed warnings from `ExtensionsGeneral.cs` file - Extracted `ProcessExtensions` class to separate file - Return null when non windows computer attempts to retrieve parent process - Updated VSCode tasks
