+++ .vscode/tasks.json
@@ -2,41 +2,58 @@
     "version": "2.0.0",
     "tasks": [
         {
-            "label": "build",
-            "command": "dotnet",
-            "type": "process",
-            "args": [
-                "build",
-                "${workspaceFolder}/src/AnalysisPrograms/AnalysisPrograms.csproj",
-                "/property:GenerateFullPaths=true",
-                "/consoleloggerparameters:NoSummary"
-            ],
-            "problemMatcher": "$msCompile"
+          "label": "build",
+          "command": "dotnet",
+          "type": "process",
+          "args": [
+            "build",
+            "${workspaceFolder}/src/AnalysisPrograms/AnalysisPrograms.csproj",
+            "/property:GenerateFullPaths=true",
+            "/consoleloggerparameters:NoSummary"
+          ],
+          "problemMatcher": "$msCompile",
+          "group": {
+            "kind": "build",
+            "isDefault": true
+          }
         },
         {
-            "label": "publish",
-            "command": "dotnet",
-            "type": "process",
-            "args": [
-                "publish",
-                "${workspaceFolder}/src/AnalysisPrograms/AnalysisPrograms.csproj",
-                "/property:GenerateFullPaths=true",
-                "/consoleloggerparameters:NoSummary"
-            ],
-            "problemMatcher": "$msCompile"
+          "label": "test",
+          "command": "dotnet",
+          "type": "process",
+          "args": [
+            "test"
+          ],
+          "problemMatcher": "$msCompile",
+          "group": {
+            "kind": "test",
+            "isDefault": true
+          }
         },
         {
-            "label": "watch",
-            "command": "dotnet",
-            "type": "process",
-            "args": [
-                "watch",
-                "run",
-                "${workspaceFolder}/src/AnalysisPrograms/AnalysisPrograms.csproj",
-                "/property:GenerateFullPaths=true",
-                "/consoleloggerparameters:NoSummary"
-            ],
-            "problemMatcher": "$msCompile"
+          "label": "publish",
+          "command": "dotnet",
+          "type": "process",
+          "args": [
+            "publish",
+            "${workspaceFolder}/src/AnalysisPrograms/AnalysisPrograms.csproj",
+            "/property:GenerateFullPaths=true",
+            "/consoleloggerparameters:NoSummary"
+          ],
+          "problemMatcher": "$msCompile"
+        },
+        {
+          "label": "watch",
+          "command": "dotnet",
+          "type": "process",
+          "args": [
+            "watch",
+            "run",
+            "${workspaceFolder}/src/AnalysisPrograms/AnalysisPrograms.csproj",
+            "/property:GenerateFullPaths=true",
+            "/consoleloggerparameters:NoSummary"
+          ],
+          "problemMatcher": "$msCompile"
         }
     ]
 }
